。

+++ main/FYP.py
@@ -46,7 +46,8 @@ def arima_invert_scaling(scaled_predictions):
     reintegrated_forecast = np.cumsum(np.insert(scaled_predictions, 0, last_known_value))[-24:]
     scaled_predictions = np.array(reintegrated_forecast).reshape(-1, 1)
     original_scale_predictions = arima_scaler.inverse_transform(scaled_predictions)
-    return float(original_scale_predictions)
+    original_scale_predictions = original_scale_predictions.astype(float)
+    return original_scale_predictions
 
 def arima_predict(input_df):
     # Check if 'month' is a column, convert it to datetime, and set it as the index
